+++ src/modules/demux/options_parser_demux.cpp
@@ -89,12 +89,20 @@ bool options_parser_demux::parse( int argc, char ***argv, options *opts )
         )
         ( "aa_counts,a", po::value<std::string>( &opts_demux->aggregate_fname )->default_value( "" ),
           "The name of the file to write aggregated counts to when aa sequences from a designed library "
-          "have multiple different nt encodings. If this option is included, names of sequences in the file supplied "
+          "have multiple different nt encodings. "
+          "If this option is included without the 'translate_aggregates' flag, names of sequences in the file supplied "
           "by the --library flag MUST be of the form ID-NUM, where ID can contain any characters but '-', and NUM "
           "represents the id of this encoding. ID and NUM MUST be separated by a single dash '-' character. If supplied "
           "aggregated counts for each sequence will be written to this file. For example, suppose we have TG1_1-1 and TG1_1-2 "
           "in our library, which says that we generated two encodings for TG1_1. If included, the file will have a single TG1_1 "
           "entry, where the count in column i is the sum of the value of column i from TG1_1-1 and TG1_1-2.\n"
+
+        )
+        ( "translate_aggregates", po::bool_switch( &opts_demux->translation_aggregation )->default_value( false ),
+          "Include this flag to use translation-based aggregation. "
+          "In this mode, nucleotide sequences that translate into the same AA sequence are considered to have been "
+          "made from the same AA sequence. "
+          "Note: When this mode is used, the name of the aggregate sequence will be the sequence that was a result of the translation.\n"
         )
         ( "samplelist,s", po::value<std::string>( &opts_demux->samplelist_fname )->required(), "A tab-delimited list of samples, one sample per line. If the samples are "
           "already indexed by I2 only the forward index (I1) and the sample name are required. The first item in each tab-delimited line is the forward (I1) index, the second "