+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- 2020-02-24: Added support for reference-independent demultiplexing.
+- 2020-02-25: [Demux] Added support for translation-based count aggregation.
+- 2020-02-24: [Demux] Added support for reference-independent demultiplexing.
 ## [1.0.0] - 2020-02-14
 Release of the first version of PepSIRF.